% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki}
\alias{aki}
\alias{aki.numeric}
\alias{aki.default}
\title{Codify AKI from Serum Creatinine and/or Urine Output}
\usage{
aki(...)

\method{aki}{numeric}(x)

\method{aki}{default}(x)
}
\arguments{
\item{x}{numeric number}
}
\value{

}
\description{
Some additional details
}
