% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aki.R
\docType{data}
\name{aki_pt_data}
\alias{aki_pt_data}
\title{AKI Patient Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 7 columns.
}
\usage{
aki_pt_data
}
\description{
A sample dataset to demonstrate calculating AKI with \code{epocakir}
}
\examples{
aki_pt_data
}
\keyword{datasets}
