% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki_SCr}
\alias{aki_SCr}
\alias{aki_SCr.default}
\alias{aki_SCr.units}
\alias{aki_SCr.numeric}
\title{AKI Staging based on changes in creatinine}
\usage{
aki_SCr(...)

\method{aki_SCr}{default}(.data, SCr, dttm, pt_id, ...)

\method{aki_SCr}{units}(SCr, dttm, pt_id, ...)

\method{aki_SCr}{numeric}(SCr, dttm, pt_id, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{dttm}{DateTime
column name, or vector if \code{.data} not provided}

\item{pt_id}{Patient ID
column name, or vector if \code{.data} not provided}
}
\value{
(ordered factor) AKI stages
}
\description{
AKI Staging based on changes in creatinine
}
\examples{
print("todo")
}
