% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\docType{data}
\name{aki_stages}
\alias{aki_stages}
\title{AKI Stages}
\format{
An object of class \code{ordered} (inherits from \code{factor}) of length 4.
}
\usage{
aki_stages
}
\description{
Ordered factor of AKI stages
}
\examples{
aki_stages
}
\keyword{datasets}
