% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki_staging}
\alias{aki_staging}
\alias{aki_staging.default}
\alias{aki_staging.units}
\alias{aki_staging.numeric}
\title{Codify AKI from Serum Creatinine and/or Urine Output}
\usage{
aki_staging(...)

\method{aki_staging}{default}(
  .data,
  SCr = NULL,
  bCr = NULL,
  UO = NULL,
  dttm = NULL,
  pt_id = NULL,
  ...
)

\method{aki_staging}{units}(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)

\method{aki_staging}{numeric}(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{bCr}{Baseline creatinine
column name, or vector if \code{.data} not provided}

\item{UO}{Urine output
column name, or vector if \code{.data} not provided}

\item{dttm}{DateTime
column name, or vector if \code{.data} not provided}

\item{pt_id}{Patient ID
column name, or vector if \code{.data} not provided}
}
\description{
Using KDIGO Clinical Practice Guideline for Acute Kidney Injury
Volume 2 | Issue 1 | March 2012
}
\details{
Provided a series of Serum Creatinine readings and/or Urine Output,
#' aki_staging() calculates whether or not a patient has AKI.
The staging (1, 2, 3) of AKI is returned
}
\examples{
print("todo")
}
