% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki_UO}
\alias{aki_UO}
\alias{aki_UO.default}
\alias{aki_UO.units}
\alias{aki_UO.numeric}
\title{AKI Staging based on urine output}
\usage{
aki_UO(...)

\method{aki_UO}{default}(.data, UO, dttm, pt_id, ...)

\method{aki_UO}{units}(UO, dttm, pt_id, ...)

\method{aki_UO}{numeric}(UO, dttm, pt_id, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{UO}{Urine output
column name, or vector if \code{.data} not provided}

\item{dttm}{DateTime
column name, or vector if \code{.data} not provided}

\item{pt_id}{Patient ID
column name, or vector if \code{.data} not provided}
}
\value{
(ordered factor) AKI stages
}
\description{
AKI Staging based on urine output
}
\examples{
print("todo")
}
