% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{Albuminuria_levels_ACR}
\alias{Albuminuria_levels_ACR}
\alias{Albuminuria_levels_ACR.default}
\alias{Albuminuria_levels_ACR.units}
\alias{Albuminuria_levels_ACR.numeric}
\title{Albuminuria Staging based on ACR}
\usage{
Albuminuria_levels_ACR(...)

\method{Albuminuria_levels_ACR}{default}(.data, ACR, ...)

\method{Albuminuria_levels_ACR}{units}(ACR, ...)

\method{Albuminuria_levels_ACR}{numeric}(ACR, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{ACR}{Albumin-to-creatinine ratio
column name, or vector if \code{.data} not provided}
}
\value{
Albuminuria category
}
\description{
Albuminuria Staging based on ACR
}
\examples{
print("todo")
}
