% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\docType{data}
\name{Albuminuria_stages}
\alias{Albuminuria_stages}
\title{Albuminuria Stages}
\format{
An object of class \code{ordered} (inherits from \code{factor}) of length 3.
}
\usage{
Albuminuria_stages
}
\description{
Ordered factor of Albuminuria stages
}
\examples{
Albuminuria_stages
}
\keyword{datasets}
