% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{Albuminuria_staging_AER}
\alias{Albuminuria_staging_AER}
\alias{Albuminuria_staging_AER.default}
\alias{Albuminuria_staging_AER.units}
\alias{Albuminuria_staging_AER.numeric}
\title{Albuminuria Staging based on AER}
\usage{
Albuminuria_staging_AER(...)

\method{Albuminuria_staging_AER}{default}(.data, AER, ...)

\method{Albuminuria_staging_AER}{units}(AER, ...)

\method{Albuminuria_staging_AER}{numeric}(AER, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{AER}{Albumin excretion rate
column name, or vector if \code{.data} not provided}
}
\value{
Albuminuria category
}
\description{
Albuminuria Staging based on AER
}
\examples{
print("todo")
}
