% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anemia.R
\name{anemia}
\alias{anemia}
\alias{anemia.numeric}
\alias{anemia.units}
\alias{anemia.default}
\title{Diagnosis of anemia from Hb concentration}
\usage{
anemia(...)

\method{anemia}{numeric}(Hb, age = NULL, birthday = NULL, gender, ...)

\method{anemia}{units}(Hb, age = NULL, birthday = NULL, gender, ...)

\method{anemia}{default}(data = NULL, age = NULL, birthday = NULL, gender, Hb, ...)
}
\arguments{
\item{...}{Further optional arguments that will be passed to method.}

\item{Hb}{The variable name e.g. "Hb" or X}

\item{age}{The variable name, e.g. "age" or X}

\item{birthday}{The variable name, e.g.  "birthday" or X}

\item{gender}{The variable name, e.g. "gender" or X}

\item{data}{A data.frame X}
}
\value{
A dataframe or X
}
\description{
TODO Make consistent with other files
}
\examples{
anemia(1:10)
}
