% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-anemia.R
\docType{data}
\name{anemia_pt_data}
\alias{anemia_pt_data}
\title{Anemia Patient Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
anemia_pt_data
}
\description{
A sample dataset to demonstrate calculating anemia with \code{epocakir}
}
\examples{
anemia_pt_data
}
\keyword{datasets}
