% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clinical_obs.R
\docType{data}
\name{clinical_obvs}
\alias{clinical_obvs}
\title{Clinical Patient Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 7 columns.
}
\usage{
clinical_obvs
}
\description{
A sample dataset to demonstrate ulitiy functions in \code{epocakir}
}
\examples{
clinical_obvs
}
\keyword{datasets}
