% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{conversion_factors}
\alias{conversion_factors}
\title{Conversion Factors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
conversion_factors
}
\description{
List of conversion factors based on tables in  KDIGO Clinical Practice
Guidelines.
}
\details{
\describe{
\item{parameter}{Name of the measurement}
\item{metric_units}{Metric units for the parameter}
\item{mol_weight}{Molecular weight (where required)}
\item{description}{Full name}
}
}
\examples{
epocakir:::conversion_factors
}
\keyword{datasets}
