% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR_adult_SCr}
\alias{eGFR_adult_SCr}
\alias{eGFR_adult_SCr.default}
\alias{eGFR_adult_SCr.units}
\alias{eGFR_adult_SCr.numeric}
\title{2009 CKD-EPI creatinine equation}
\usage{
eGFR_adult_SCr(...)

\method{eGFR_adult_SCr}{default}(.data, SCr, Age, male, black, ...)

\method{eGFR_adult_SCr}{units}(SCr, Age, male, black, ...)

\method{eGFR_adult_SCr}{numeric}(SCr, Age, male, black, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{Age}{Age of patient
column name, or vector if \code{.data} not provided}

\item{male}{Male or not
column name, or vector if \code{.data} not provided}

\item{black}{Black race or not
column name, or vector if \code{.data} not provided}
}
\value{
Estimated GFR
of the same type provided (numeric or units)
}
\description{
2009 CKD-EPI creatinine equation
}
\examples{
print("todo")
}
