% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR_adult_SCr_SCysC}
\alias{eGFR_adult_SCr_SCysC}
\alias{eGFR_adult_SCr_SCysC.default}
\alias{eGFR_adult_SCr_SCysC.units}
\alias{eGFR_adult_SCr_SCysC.numeric}
\title{2012 CKD-EPI creatinine-cystatin C equation}
\usage{
eGFR_adult_SCr_SCysC(...)

\method{eGFR_adult_SCr_SCysC}{default}(.data, SCr, SCysC, Age, male, black, ...)

\method{eGFR_adult_SCr_SCysC}{units}(SCr, SCysC, Age, male, black, ...)

\method{eGFR_adult_SCr_SCysC}{numeric}(SCr, SCysC, Age, male, black, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{SCysC}{Serum Cystatin C
column name, or vector if \code{.data} not provided}

\item{Age}{Age of patient
column name, or vector if \code{.data} not provided}

\item{male}{Male or not
column name, or vector if \code{.data} not provided}

\item{black}{Black race or not
column name, or vector if \code{.data} not provided}
}
\value{
Estimated GFR
of the same type provided (numeric or units)
}
\description{
2012 CKD-EPI creatinine-cystatin C equation
}
\examples{
print("todo")
}
