% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR_child_SCr}
\alias{eGFR_child_SCr}
\alias{eGFR_child_SCr.default}
\alias{eGFR_child_SCr.units}
\alias{eGFR_child_SCr.numeric}
\title{Pediatric SCr and Height}
\usage{
eGFR_child_SCr(...)

\method{eGFR_child_SCr}{default}(.data, SCr, height, ...)

\method{eGFR_child_SCr}{units}(SCr, height, ...)

\method{eGFR_child_SCr}{numeric}(SCr, height, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{height}{Height of patient
column name, or vector if \code{.data} not provided}
}
\value{
Estimated GFR
of the same type provided (numeric or units)
}
\description{
Pediatric SCr and Height
}
\examples{
print("todo")
}
