% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR_child_SCysC}
\alias{eGFR_child_SCysC}
\alias{eGFR_child_SCysC.default}
\alias{eGFR_child_SCysC.units}
\alias{eGFR_child_SCysC.numeric}
\title{Pediatric SCysC}
\usage{
eGFR_child_SCysC(...)

\method{eGFR_child_SCysC}{default}(.data, SCysC, ...)

\method{eGFR_child_SCysC}{units}(SCysC, ...)

\method{eGFR_child_SCysC}{numeric}(SCysC, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCysC}{Serum Cystatin C
column name, or vector if \code{.data} not provided}
}
\value{
Estimated GFR
of the same type provided (numeric or units)
}
\description{
Pediatric SCysC
}
\examples{
print("todo")
}
