% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eGFR.R
\docType{data}
\name{eGFR_pt_data}
\alias{eGFR_pt_data}
\title{eGFR Patient Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 10 columns.
}
\usage{
eGFR_pt_data
}
\description{
A sample dataset to demonstrate calculating eGFR with \code{epocakir}
}
\examples{
eGFR_pt_data
}
\keyword{datasets}
