% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{generate_cr_ch}
\alias{generate_cr_ch}
\title{Generate creatinine changes}
\usage{
generate_cr_ch(data, SCr, dttm, pt_id = NULL)
}
\arguments{
\item{data}{A data.frame of datetime and creatinine changes}

\item{SCr}{The variable name for serum creatinine column (units)}

\item{dttm}{The variable name for the datetime column (POSIXct)}

\item{pt_id}{The variable name for the patient id (optional, character)}
}
\description{
Generate creatinine changes
}
\examples{
data_ <- data.frame(
pt_id_ = c(rep("pt1", 3 + 3), rep("pt2", 3)),
dttm_ = c(
  seq(
    lubridate::as_datetime("2020-10-18 09:00:00", tz = "Australia/Melbourne"),
    lubridate::as_datetime("2020-10-20 09:00:00", tz = "Australia/Melbourne"),
    length.out = 3
  ),
  seq(
    lubridate::as_datetime("2020-10-23 09:00:00", tz = "Australia/Melbourne"),
    lubridate::as_datetime("2020-10-25 21:00:00", tz = "Australia/Melbourne"),
    length.out = 3
  ),
  seq(
    lubridate::as_datetime("2020-10-18 10:00:00", tz = "Australia/Melbourne"),
    lubridate::as_datetime("2020-10-19 10:00:00", tz = "Australia/Melbourne"),
    length.out = 3
  )
),
SCr_ = c(
  units::set_units(seq(2.0, 3.0, by = 0.5), "mg/dl"),
  units::set_units(seq(3.5, 4.0, by = 0.25), "mg/dl"),
  units::set_units(seq(3.3, 3.5, by = 0.10), "mg/dl")
),
bCr_ = c(
  rep(units::set_units(1.8, "mg/dl"), 3 + 3),
  rep(units::set_units(3.0, "mg/dl"), 3)
)
)
data <- data_[sample(nrow(data_)), ]

generate_cr_ch(data, SCr = "SCr_", dttm = "dttm_", pt_id = "pt_id_")

}
