% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\docType{data}
\name{GFR_stages}
\alias{GFR_stages}
\title{GFR Stages}
\format{
An object of class \code{ordered} (inherits from \code{factor}) of length 6.
}
\usage{
GFR_stages
}
\description{
Ordered factor of GFR stages
}
\details{
See \url{https://kdigo.org/guidelines/ckd-evaluation-and-management/} for more details
}
\examples{
GFR_stages
}
\keyword{datasets}
