% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki}
\alias{aki}
\title{Codify AKI}
\usage{
aki(sCr, bCr, UO)
}
\arguments{
\item{sCr}{numeric}

\item{bCr}{numeric}

\item{UO}{numeric}
}
\value{
numeric
}
\description{
Codify AKI
}
\examples{
aki(1, 1, 1)
}
