% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki}
\alias{aki}
\alias{aki.numeric}
\alias{aki.ts}
\alias{aki.default}
\title{Codify AKI from Serum Creatinine and/or Urine Output}
\usage{
aki(x, ...)

\method{aki}{numeric}(SCr, bCr = NULL, na.rm = FALSE)

\method{aki}{ts}(ts_data)

\method{aki}{default}(x)
}
\arguments{
\item{x}{numeric number}
}
\value{

}
\description{
Some additional details
}
