% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aki.R
\name{aki}
\alias{aki}
\alias{aki.default}
\alias{aki.units}
\alias{aki.numeric}
\title{Codify AKI from Serum Creatinine and/or Urine Output}
\usage{
aki(...)

\method{aki}{default}(.data, SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)

\method{aki}{units}(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)

\method{aki}{numeric}(SCr = NULL, bCr = NULL, UO = NULL, dttm = NULL, pt_id = NULL, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{bCr}{Baseline creatinine
column name, or vector if \code{.data} not provided}

\item{UO}{Urine output
column name, or vector if \code{.data} not provided}

\item{dttm}{DateTime
column name, or vector if \code{.data} not provided}

\item{pt_id}{Patient ID
column name, or vector if \code{.data} not provided}
}
\description{
Using KDIGO Clinical Practice Guideline for Acute Kidney Injury
Volume 2 | Issue 1 | March 2012
}
\details{
Provided a series of Serum Creatinine readings and/or Urine Output, aki()
calculates whether or not a patient has AKI. The staging (1, 2, 3) of AKI is
also calculated
}
\examples{
print("todo")
}
