% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\docType{data}
\name{Albuminuria_stages}
\alias{Albuminuria_stages}
\title{Albuminuria Stages}
\format{
An object of class \code{ordered} (inherits from \code{factor}) of length 4.
}
\usage{
Albuminuria_stages
}
\description{
Ordered factor of Albuminuria stages
}
\details{
\itemize{
\item{A1: Normal to mildly increased}
\item{A2: Moderately increased}
\item{A3: Severely increased}
}

See \url{https://kdigo.org/guidelines/ckd-evaluation-and-management/} for more details
}
\examples{
Albuminuria_stages
}
\keyword{datasets}
