% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anemia.R
\name{anemia}
\alias{anemia}
\alias{anemia.default}
\alias{anemia.units}
\alias{anemia.numeric}
\title{Diagnosis of anemia from Hb concentration}
\usage{
anemia(...)

\method{anemia}{default}(.data, Hb, age, male, ...)

\method{anemia}{units}(Hb, age, male, ...)

\method{anemia}{numeric}(Hb, age, male, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{Hb}{Hemoglobin concenration
column name, or vector if \code{.data} not provided}

\item{age}{Age of patient
column name, or vector if \code{.data} not provided}

\item{male}{Male or not
column name, or vector if \code{.data} not provided}
}
\value{
Anemia
as logical \code{TRUE} or \code{FALSE}
}
\description{
Diagnosis of anemia from Hb concentration
}
\examples{
print("todo")
}
