% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_date_time_cols}
\alias{combine_date_time_cols}
\title{Combine date and time columns into a single DateTime column}
\usage{
combine_date_time_cols(.data)
}
\arguments{
\item{.data}{(data.frame) A data frame or data frame extension (e.g. a tibble)}
}
\value{
(data.frame) An object of the same type as \code{.data}
}
\description{
Combine date and time columns into a single DateTime column
}
\examples{
print("todo")
}
