% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combn_changes}
\alias{combn_changes}
\alias{combn_changes.default}
\alias{combn_changes.POSIXct}
\title{Combinatorics changes}
\usage{
combn_changes(...)

\method{combn_changes}{default}(.data, dttm, col, pt_id = NULL, ...)

\method{combn_changes}{POSIXct}(dttm, col, pt_id, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{dttm}{DateTime
column name, or vector if \code{.data} not provided}

\item{col}{Variable
column name, or vector if \code{.data} not provided}

\item{pt_id}{Patient ID
column name, or vector if \code{.data} not provided}
}
\value{
Combinatorics changes
of the same type provided (numeric or units)
}
\description{
Combinatorics changes
}
\examples{
print("todo")
}
