% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR}
\alias{eGFR}
\alias{eGFR.default}
\alias{eGFR.units}
\alias{eGFR.numeric}
\title{GFR Estimation}
\usage{
eGFR(...)

\method{eGFR}{default}(
  .data,
  SCr = NULL,
  SCysC = NULL,
  Age = NULL,
  height = NULL,
  BUN = NULL,
  male = NULL,
  black = NULL,
  pediatric = NULL,
  ...
)

\method{eGFR}{units}(
  SCr = NULL,
  SCysC = NULL,
  Age = NULL,
  height = NULL,
  BUN = NULL,
  male = NULL,
  black = NULL,
  pediatric = NULL,
  ...
)

\method{eGFR}{numeric}(
  SCr = NULL,
  SCysC = NULL,
  Age = NULL,
  height = NULL,
  BUN = NULL,
  male = NULL,
  black = NULL,
  pediatric = NULL,
  ...
)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCr}{Serum creatinine
column name, or vector if \code{.data} not provided}

\item{SCysC}{Serum Cystatin C
column name, or vector if \code{.data} not provided}

\item{Age}{Age of patient
column name, or vector if \code{.data} not provided}

\item{height}{Height of patient
column name, or vector if \code{.data} not provided}

\item{BUN}{Blood urea nitrogen
column name, or vector if \code{.data} not provided}

\item{male}{Male or not
column name, or vector if \code{.data} not provided}

\item{black}{Black race or not
column name, or vector if \code{.data} not provided}

\item{pediatric}{(logical) Paediatric or not
column name, or vector if \code{.data} not provided}
}
\value{
(units) Estimated glomerular filtration rate (eGFR)
of the same type provided (numeric or units in ml/min/1.73m2)
}
\description{
\itemize{
\item 2009 CKD-EPI creatinine equation
\item 2012 CKD-EPI cystatin C equation
\item 2012 CKD-EPI creatinine-cystatin C equation
\item Pediatric equations
}
}
\examples{
print("todo")
}
