% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR}
\alias{eGFR}
\alias{eGFR_adult_SCysC}
\alias{eGFR_adult_SCr_SCysC}
\alias{eGFR_child_SCr}
\alias{eGFR_child_SCr_BUN}
\alias{eGFR_child_SCysC}
\title{GFR Estimation}
\usage{
eGFR(
  SCr = NULL,
  SCysC = NULL,
  Age = NULL,
  height = NULL,
  BUN = NULL,
  male = FALSE,
  black = FALSE,
  pediatric = FALSE
)

eGFR_adult_SCysC(SCysC, Age, male)

eGFR_adult_SCr_SCysC(SCr, SCysC, Age, male, black)

eGFR_child_SCr(SCr, height)

eGFR_child_SCr_BUN(SCr, height, BUN)

eGFR_child_SCysC(SCysC)
}
\arguments{
\item{SCr}{(units) Serum creatinine, converted to mg/dl}

\item{SCysC}{(units) Serum cystatin C, converted to mg/l}

\item{Age}{(units) Age of patient, converted to years}

\item{height}{(units) Height of patient, converted to m}

\item{BUN}{(units) Blood urea nitrogen, converted to mg/dl}

\item{male}{(logical) Male or not}

\item{black}{(logical) Black race or not}

\item{pediatric}{(logical) Paediatric or not}
}
\value{
(units) Estimated glomerular filtration rate (eGFR) in ml/min/1.73m2
}
\description{
\itemize{
\item 2009 CKD-EPI creatinine equation
\item 2012 CKD-EPI cystatin C equation
\item 2012 CKD-EPI creatinine-cystatin C equation
\item Pediatric equations
}
}
\examples{
print("todo")
}
