% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{eGFR_adult_SCysC}
\alias{eGFR_adult_SCysC}
\alias{eGFR_adult_SCysC.default}
\alias{eGFR_adult_SCysC.units}
\alias{eGFR_adult_SCysC.numeric}
\title{2012 CKD-EPI cystatin C equation}
\usage{
eGFR_adult_SCysC(...)

\method{eGFR_adult_SCysC}{default}(.data, SCysC, Age, male, ...)

\method{eGFR_adult_SCysC}{units}(SCysC, Age, male, ...)

\method{eGFR_adult_SCysC}{numeric}(SCysC, Age, male, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{SCysC}{Serum Cystatin C
column name, or vector if \code{.data} not provided}

\item{Age}{Age of patient
column name, or vector if \code{.data} not provided}

\item{male}{Male or not
column name, or vector if \code{.data} not provided}
}
\value{
Estimated GFR
of the same type provided (numeric or units)
}
\description{
2012 CKD-EPI cystatin C equation
}
\examples{
print("todo")
}
