% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd.R
\name{GFR_staging}
\alias{GFR_staging}
\alias{GFR_staging.default}
\alias{GFR_staging.units}
\alias{GFR_staging.numeric}
\title{GFR Staging}
\usage{
GFR_staging(...)

\method{GFR_staging}{default}(.data, GFR, ...)

\method{GFR_staging}{units}(GFR, ...)

\method{GFR_staging}{numeric}(GFR, ...)
}
\arguments{
\item{...}{Further optional arguments}

\item{.data}{(data.frame) A data.frame, optional}

\item{GFR}{Glomerular filtration rate
column name, or vector if \code{.data} not provided}
}
\value{
GFR category
}
\description{
GFR Staging
}
\examples{
print("todo")
}
